Clean up goals redux to match others
