+++ src/components/GoalTimeline/tests/index.test.tsx
@@ -7,8 +7,11 @@ import "tests/reactI18nextMock";
 
 import { Permission } from "api/models";
 import GoalTimeline, { createSuggestionData } from "components/GoalTimeline";
-import { defaultState } from "components/GoalTimeline/DefaultState";
-import { Goal, GoalType, GoalsState } from "types/goals";
+import {
+  type GoalsState,
+  defaultState,
+} from "components/GoalTimeline/Redux/GoalReduxTypes";
+import { Goal, GoalType } from "types/goals";
 import { goalTypeToGoal } from "utilities/goalUtilities";
 
 jest.mock("backend", () => ({
